refactor: move versioninfo package to source directory
