+++ webapi/src/main/scala/org/knora/webapi/app/BUILD.bazel
@@ -18,7 +18,6 @@ scala_library(
         "//webapi/src/main/scala/org/knora/webapi/routing",
         "//webapi/src/main/scala/org/knora/webapi/settings",
         "//webapi/src/main/scala/org/knora/webapi/store",
-        "//webapi/src/main/scala/org/knora/webapi/util",
         "//webapi/src/main/scala/org/knora/webapi/util/cache",
         "@maven//:ch_megard_akka_http_cors_2_12",
         "@maven//:com_github_swagger_akka_http_swagger_akka_http_2_12",
