+++ webapi/src/main/scala/org/knora/webapi/http/version/ServerVersion.scala
@@ -22,6 +22,7 @@ package org.knora.webapi.http.version
 import akka.http.scaladsl.model.headers.Server
 import akka.http.scaladsl.server.Directives.respondWithHeader
 import akka.http.scaladsl.server.Route
+import org.knora.webapi.http.version.versioninfo.VersionInfo
 
 /**
  * This object provides methods that can be used to add the [[Server]] header
@@ -38,6 +39,4 @@ object ServerVersion {
     def addServerHeader(route: Route): Route = respondWithHeader(serverVersionHeader) {
         route
     }
-
-
 }
