+++ webapi/src/main/scala/org/knora/webapi/http/version/versioninfo/BUILD.bazel
@@ -21,7 +21,7 @@ stamp(
 # Generates a scala source file that includes version information
 # The generated intermediate file is missing the BUILD_TAG, which is
 # added by the 'version_info_with_build_tag' genrule
-load("//tools/version_info:gen_version_info.bzl", "gen_version_info")
+load("//webapi/src/main/scala/org/knora/webapi/http/version/versioninfo:gen_version_info.bzl", "gen_version_info")
 gen_version_info(
     name = "version_info_without_build_tag",
     scala_version = SCALA_VERSION,
@@ -44,6 +44,6 @@ genrule(
 )
 
 scala_library(
-    name = "version_info",
+    name = "versioninfo",
     srcs = [":version_info_with_build_tag"],
 )
