+++ webapi/src/main/scala/org/knora/webapi/http/version/versioninfo/add_build_tag.sh
