+++ webapi/src/main/scala/org/knora/webapi/messages/BUILD.bazel
@@ -13,6 +13,7 @@ scala_library(
         "//webapi/src/main/scala/org/knora/webapi/settings",
         "//webapi/src/main/scala/org/knora/webapi/util",
         "//webapi/src/main/scala/org/knora/webapi/util/cache",
+        "@maven//:com_apicatalog_titanium_json_ld",
         "@maven//:com_google_gwt_gwt_servlet",
         "@maven//:com_ibm_icu_icu4j",
         "@maven//:com_sksamuel_diff_diff",
@@ -20,19 +21,17 @@ scala_library(
         "@maven//:com_typesafe_akka_akka_http_2_12",
         "@maven//:com_typesafe_akka_akka_http_core_2_12",
         "@maven//:com_typesafe_akka_akka_http_spray_json_2_12",
-        "@maven//:com_apicatalog_titanium_json_ld",
-        "@maven//:javax_json_javax_json_api",
-        "@maven//:org_glassfish_jakarta_json",
         "@maven//:com_typesafe_akka_akka_stream_2_12",
         "@maven//:com_typesafe_play_twirl_api_2_12",
         "@maven//:com_typesafe_scala_logging_scala_logging_2_12",
-        "@maven//:org_apache_jena_apache_jena_libs",
         "@maven//:commons_io_commons_io",
         "@maven//:commons_validator_commons_validator",
         "@maven//:io_spray_spray_json_2_12",
+        "@maven//:javax_json_javax_json_api",
         "@maven//:net_sf_saxon_Saxon_HE",
         "@maven//:org_apache_commons_commons_lang3",
         "@maven//:org_apache_commons_commons_text",
+        "@maven//:org_apache_jena_apache_jena_libs",
         "@maven//:org_eclipse_rdf4j_rdf4j_client",
         "@maven//:org_jodd_jodd",
         "@maven//:org_scala_lang_modules_scala_java8_compat_2_12",
