+++ webapi/src/main/scala/org/knora/webapi/routing/BUILD.bazel
@@ -7,11 +7,11 @@ scala_library(
     srcs = glob(["**/*.scala"]),
     unused_dependency_checker_mode = "warn",
     deps = [
-        "//tools/version_info",
         "//webapi/src/main/scala/org/knora/webapi",
         "//webapi/src/main/scala/org/knora/webapi/annotation",
         "//webapi/src/main/scala/org/knora/webapi/exceptions",
         "//webapi/src/main/scala/org/knora/webapi/http/status",
+        "//webapi/src/main/scala/org/knora/webapi/http/version/versioninfo",
         "//webapi/src/main/scala/org/knora/webapi/instrumentation",
         "//webapi/src/main/scala/org/knora/webapi/messages",
         "//webapi/src/main/scala/org/knora/webapi/settings",
@@ -30,14 +30,13 @@ scala_library(
         "@maven//:com_typesafe_config",
         "@maven//:com_typesafe_play_twirl_api_2_12",
         "@maven//:com_typesafe_scala_logging_scala_logging_2_12",
-        "@maven//:org_apache_jena_apache_jena_libs",
         "@maven//:commons_validator_commons_validator",
         "@maven//:io_spray_spray_json_2_12",
         "@maven//:io_swagger_swagger_annotations",
         "@maven//:io_swagger_swagger_jaxrs",
         "@maven//:io_swagger_swagger_models",
         "@maven//:javax_ws_rs_jsr311_api",
-        "@maven//:org_eclipse_rdf4j_rdf4j_client",
+        "@maven//:org_apache_jena_apache_jena_libs",
         "@maven//:org_scala_lang_modules_scala_xml_2_12",
         "@maven//:org_slf4j_slf4j_api",
     ],
