+++ webapi/src/main/scala/org/knora/webapi/util/BUILD.bazel
@@ -12,7 +12,6 @@ scala_library(
         "@maven//:com_jsuereth_scala_arm_2_12",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
         "@maven//:com_typesafe_akka_akka_http_core_2_12",
-        "@maven//:com_typesafe_akka_akka_stream_2_12",
         "@maven//:com_typesafe_scala_logging_scala_logging_2_12",
         "@maven//:commons_validator_commons_validator",
         "@maven//:io_spray_spray_json_2_12",
