+++ webapi/src/test/scala/org/knora/webapi/http/BUILD.bazel
@@ -1,0 +1,1 @@
+package(default_visibility = ["//visibility:public"])
