+++ webapi/src/test/scala/org/knora/webapi/http/version/versioninfo/BUILD.bazel
@@ -4,10 +4,10 @@ load("@io_bazel_rules_scala//scala:scala.bzl", "scala_test")
 load("//third_party:dependencies.bzl", "ALL_WEBAPI_MAIN_DEPENDENCIES", "BASE_TEST_DEPENDENCIES", "BASE_TEST_DEPENDENCIES_WITH_JSON", "BASE_TEST_DEPENDENCIES_WITH_JSON_LD")
 
 scala_test(
-    name = "ServerVersionSpec",
+    name = "VersionInfoSpec",
     size = "small",  # 60s
     srcs = [
-        "ServerVersionSpec.scala",
+        "VersionInfoSpec.scala",
     ],
     jvm_flags = ["-Dconfig.resource=fuseki.conf"],
     # unused_dependency_checker_mode = "warn",
@@ -17,8 +17,7 @@ scala_test(
         "@maven//:com_typesafe_akka_akka_slf4j_2_12",
         "@maven//:org_slf4j_log4j_over_slf4j",
     ],
-    deps = ALL_WEBAPI_MAIN_DEPENDENCIES + [
-        "//webapi:main_library",
-        "//webapi:test_library",
+    deps = [
+        "//webapi/src/main/scala/org/knora/webapi/http/version/versioninfo",
     ] + BASE_TEST_DEPENDENCIES_WITH_JSON,
 )
