+++ webapi/src/test/scala/org/knora/webapi/http/version/versioninfo/VersionInfoSpec.scala
@@ -17,9 +17,8 @@
  *  License along with Knora.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package webapi.src.test.scala.org.knora.webapi.http.version
+package org.knora.webapi.http.version.versioninfo
 
-import org.knora.webapi.http.version.VersionInfo
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpecLike
